+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project follows [these versioning rules](https://sagegerard.com/edition-rev
 - Add `mkint` command
 - Help authors make input expressions
 - Redirected downloads with no limit now work properly
+- Distinguish between abstract and concrete inputs
 - Add missing flags to CLI
 
 ## [draft:alpha] - 2020-08-02